# contracts/analytics.openapi.yaml
openapi: 3.0.3
info:
  title: FIT4110 B5 - Analytics Service API
  version: 1.0.0
  description: |
    Analytics Service của nhóm B5 dùng để tổng hợp dữ liệu,
    thống kê và tích hợp với các service khác.
servers:
  - url: http://analytics_api:8000
    description: Docker Compose (class-net)
  - url: http://localhost:8000
    description: Local Development
...
paths:
...
  /external/iot/mock-check:
    post:
      tags:
        - Integrations
      summary: Đẩy thử dữ liệu mô phỏng sang IoT Ingestion (B1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IoTDataPayload'
      responses:
        "200":
          description: IoT Ingestion nhận dữ liệu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Data ingested successfully"
        "400":
          description: Dữ liệu gửi lên sai cấu trúc
        "503":
          description: Cannot connect to IoT service
# B3: Access Gate - Bảo
...
  /external/access/check:
    post:
      tags:
        - Integrations
      summary: Gọi Access Gate để kiểm tra quyền truy cập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessCheckRequest'
      responses:
        "200":
          description: Access Gate trả về kết quả hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessCheckResponse'
        "400":
          description: Thiếu thông tin check-in (như mã thẻ hoặc ID cửa)
        "503":
          description: Cannot connect to Access Gate service
...
components:
  schemas:
...
    IoTDataPayload:
      type: object
      required:
        - device_id
        - temperature
        - humidity
      properties:
        device_id:
          type: string
          example: "SENSOR-DHT22-01"
        temperature:
          type: number
          format: float
          example: 28.5
        humidity:
          type: number
          format: float
          example: 60.0
        timestamp:
          type: string
          format: date-time
          example: "2026-06-09T13:45:00Z"

    # 2. Định nghĩa dữ liệu yêu cầu kiểm tra cửa (B3)
    AccessCheckRequest:
      type: object
      required:
        - card_uid
        - gate_id
      properties:
        card_uid:
          type: string
          example: "A3B2C1D0"
        gate_id:
          type: string
          example: "GATE-MAIN-01"

    # 3. Định nghĩa dữ liệu trả về từ cửa (B3)
    AccessCheckResponse:
      type: object
      properties:
        is_allowed:
          type: boolean
          example: true
        employee_name:
          type: string
          example: "Nguyen Van A"
        timestamp:
          type: string
          format: date-time
          example: "2026-06-09T13:46:12Z"

    # 4. Định nghĩa dữ liệu nhận từ Camera (B2) - Dùng khi B2 hoàn thiện
...